pyproject.toml
[project]
name = "realign-templates"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.34.0",
    "openpyxl>=3.1.5",
    "ibm-watsonx-ai>=0.0.5",
    "pandas<2.3.0",
    "python-dotenv>=1.1.0",
]
